新增用户名重复验证、角色名重复验证

+++ urms_web/src/main/java/com/urms/service/impl/RoleServiceImpl.java
@@ -4,7 +4,10 @@ import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
 
 import com.urms.entity.QueryRoleCondition;
 import com.urms.entity.Role;
+import com.urms.handle.BusinessException;
 import com.urms.mapper.RoleMapper;
+import com.urms.mapper.UserRoleMapper;
+import com.urms.response.ResultCode;
 import com.urms.service.RoleService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -23,6 +26,8 @@ import java.util.List;
 public class RoleServiceImpl extends ServiceImpl<RoleMapper, Role> implements RoleService {
     @Autowired
     private RoleMapper roleMapper;
+    @Autowired
+    private UserRoleMapper userRoleMapper;
 
 
     @Override
@@ -44,11 +49,19 @@ public class RoleServiceImpl extends ServiceImpl<RoleMapper, Role> implements Ro
 
     @Override
     public int insertRole(Role role) {
+        Role selectByRoleName = roleMapper.selectByRoleName(role.getRoleName());
+        if (selectByRoleName!=null){
+            throw new BusinessException(ResultCode.ROLE_AlREADY_EXISTS_EXCEPTION.getCode(), "角色名已存在");
+        }
         return roleMapper.insertRole(role);
     }
 
     @Override
     public int deleteRoleById(Integer roleId) {
+        List<Integer> userIds = userRoleMapper.selectUserIds(roleId);
+        if (userIds!=null){
+            throw new BusinessException(ResultCode.ROLE_ASSIGNED_USER_EXCEPTION.getCode(), "角色已分配给用户");
+        }
         return roleMapper.deleteRoleById(roleId);
     }
